requirements.txt
streamlit
crewai
crewai-tools
python-dotenv
litellm
langchain-google-genai
pyyaml